Alterações no projeto!

+++ blogPessoal/src/main/java/org/generation/blogPessoal/controller/PostagemController.java
@@ -1,6 +1,9 @@
 package org.generation.blogPessoal.controller;
 
 import java.util.List;
+import java.util.Optional;
+
+import javax.validation.Valid;
 
 import org.generation.blogPessoal.model.Postagem;
 import org.generation.blogPessoal.repository.PostagemRepository;
@@ -16,11 +19,13 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.server.ResponseStatusException;
 
 @RestController
 @RequestMapping("/postagens")	//vai aceita coisas de postagem
-@CrossOrigin("*")		//aceita coisa de outras origens como react e afins
+@CrossOrigin(value = "*", allowedHeaders = "*")		//aceita coisa de outras origens como react e afins
 public class PostagemController {
 	
 	
@@ -32,6 +37,7 @@ public class PostagemController {
 		
 		return ResponseEntity.ok(repository.findAll());
 	}
+	
 	
 	@GetMapping("/{id}")		//metodo de busca por id no BD
 	public ResponseEntity<Postagem> GetById(@PathVariable long id){
@@ -47,19 +53,28 @@ public class PostagemController {
 	}
 	
 	@PostMapping		//inserir dados na tabela via body
-	public ResponseEntity<Postagem> post (@RequestBody Postagem postagem){
-		return ResponseEntity.status(HttpStatus.CREATED).body(repository.save(postagem));
+	public ResponseEntity<Postagem> post (@Valid @RequestBody Postagem postagem){
+		return ResponseEntity.status(HttpStatus.CREATED)
+							.body(repository.save(postagem));
 		
 	}
 	
 	@PutMapping			// alterar dados na tabela via body
-	public ResponseEntity<Postagem> put (@RequestBody Postagem postagem){
+	public ResponseEntity<Postagem> put (@Valid @RequestBody Postagem postagem){
 		
-		return ResponseEntity.status(HttpStatus.OK).body(repository.save(postagem));
+		return repository.findById(postagem.getId())
+				.map(resp -> ResponseEntity.status(HttpStatus.OK)
+						.body(repository.save(postagem)))
+				.orElse(ResponseEntity.status(HttpStatus.BAD_REQUEST).build());
 	}
+	
+	@ResponseStatus(HttpStatus.NO_CONTENT)
 	@DeleteMapping ("/{id}")
 	public void delete(@PathVariable long id) {
-		repository.deleteById(id);
+		Optional<Postagem> post = repository.findById(id);
+		 if(post.isEmpty())
+			 throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
+		 repository.deleteById(id);
 	}
